+++ en/settings.js
@@ -152,5 +152,8 @@ export default {
   'settings.accessibility.ruler': 'Attach a ruler to the mouse',
 
   'settings.tab.legal': 'Legal documents',
-  'settings.tab.exportDelete': 'Export data & delete account'
+  'settings.tab.exportDelete': 'Export data & delete account',
+  'settings.languages.uiLanguage': 'Interface language',
+  'settings.languages.readingLanguage': 'Reading language',
+  'settings.languages.readingLanguageExplained': 'When looking up stories we will offer you stories and their translations in this language unless you specify otherwise.'
 }