+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added `sharing.bsky`
 * Added `common.content`
+* Added reading language keys
+* Small settings re-adjustment to better support client-side
 
 ## 2024-12-28 - v2.15.1